+++ lib/utils.php
@@ -86,20 +86,6 @@ function roots_title() {
 }
 
 /**
- * Show an admin notice if .htaccess isn't writable
- */
-function roots_htaccess_writable() {
-  if (!is_writable(get_home_path() . '.htaccess')) {
-    if (current_user_can('administrator')) {
-      add_action('admin_notices', create_function('', "echo '<div class=\"error\"><p>" . sprintf(__('Please make sure your <a href="%s">.htaccess</a> file is writable ', 'roots'), admin_url('options-permalink.php')) . "</p></div>';"));
-    }
-  }
-}
-if (current_theme_supports('h5bp-htaccess')) {
-  add_action('admin_init', 'roots_htaccess_writable');
-}
-
-/**
  * Return WordPress subdirectory if applicable
  */
 function wp_base_dir() {